Propagate refactor of IAugmentable
